Create Legend module
This contains most of the legend-related types, but not LegendConfig
