+++ hvega/hvega.cabal
@@ -128,6 +128,7 @@ library
                        Graphics.Vega.VegaLite.Foundation
                        Graphics.Vega.VegaLite.Geometry
                        Graphics.Vega.VegaLite.Input
+                       Graphics.Vega.VegaLite.Legend
                        Graphics.Vega.VegaLite.Mark
                        Graphics.Vega.VegaLite.Output
                        Graphics.Vega.VegaLite.Selection
